+++ mandatory/server.c
@@ -1,4 +1,16 @@
-# include "minitalk.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   server.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yjaafar <yjaafar@student.1337.ma>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/23 07:12:27 by yjaafar           #+#    #+#             */
+/*   Updated: 2025/03/23 07:17:53 by yjaafar          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minitalk.h"
 
 static t_client	*g_client;
 
@@ -17,7 +29,7 @@ static void	message_handler(int signum)
 		if (g_client->bit == 8 && !g_client->byte)
 		{
 			kill(g_client->client_pid, SIGUSR2);
-			delete_client(g_client);
+			delete_client(&g_client);
 		}
 	}
 	if (g_client && g_client->bit == 8)
@@ -62,7 +74,7 @@ static void	newnode(int signum, pid_t client_pid)
 	}
 }
 
-static void signal_handler(int signum, siginfo_t *info, void *context)
+static void	signal_handler(int signum, siginfo_t *info, void *context)
 {
 	(void)context;
 	if (!g_client || g_client->client_pid != info->si_pid)
@@ -85,21 +97,21 @@ int	main(void)
 {
 	struct sigaction	act;
 
-	act.sa_flags = SA_SIGINFO;
+	act.sa_flags = SA_SIGINFO | SA_RESTART;
 	display_banner();
 	if (sigemptyset(&act.sa_mask) == -1)
-		return (write(1, "sigemptyset failed\n", 19), 1);
-	sigaddset(&act.sa_mask, SIGUSR1);
-	sigaddset(&act.sa_mask, SIGUSR2);
+		return (ft_putstr("sigemptyset failed\n"), 1);
+	if (sigaddset(&act.sa_mask, SIGUSR1) == -1
+		|| sigaddset(&act.sa_mask, SIGUSR2) == -1)
+		return (ft_putstr("sigaddset failed\n"), 1);
 	act.sa_sigaction = signal_handler;
 	if (sigaction(SIGUSR1, &act, NULL) == -1
 		|| sigaction(SIGUSR2, &act, NULL) == -1)
-		return (write(1, "sigaction failed\n", 17), 1);
+		return (ft_putstr("sigaction failed\n"), 1);
 	while (1)
 	{
 		if (g_client)
 			if (kill(g_client->client_pid, 0) == -1)
-				delete_client(g_client);
-		// pause();
+				delete_client(&g_client);
 	}
 }
